codemirror設定を複数に分割
